requirements.txt
awsebcli==3.7.3
blessed==1.9.5
botocore==1.4.4
cement==2.4.0
colorama==0.3.3
Django==1.9.4
django-cors-headers==1.1.0
django-filter==0.11.0
djangorestframework==3.3.1
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
jmespath==0.9.0
Markdown==2.6.5
MySQL-python==1.2.5
pathspec==0.3.3
python-dateutil==2.5.0
PyYAML==3.11
requests==2.6.2
six==1.10.0
texttable==0.8.4
virtualenv==13.1.2
wcwidth==0.1.6
websocket-client==0.35.0
